refactor: moved error report code into separate file
